requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.15
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-environ==0.11.2
gunicorn==23.0.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
whitenoise==6.7.0
django-storages[google]
google-auth
google-cloud-storage
django-storages